build(docker): split volumes into internal and storage
When Azurite and LocalStack are mounted, they create anonymous volumes to store internal data. We can control this by setting these folders to named volumes. However, this will create a folder that mixes internal and storage files. We split the volumes and correctly set each to load internal or storage files.