refactor(aws_tools): set default to environment variables
The function os.getenv returns None if the variable does not exist. This causes mypy to raise an error about required_vars and _load_env_vars returning a typing hint. We set up os.getenv to return "" instead of None. In addition, the logger at method downlaod_file was removed.